refactor(cli): extract vault + pipeline handlers from cli.py (ARCH-2)
Move 23 handlers out of the cli.py god-file into two new sibling modules
(real-move; byte-faithful bodies, no shim):
  - cli_vault.py (10): _synthesize, _cleanup_hub, _vault_graph_colors,
    _vault_hub_backlink_migrate, _vault_install_theme,
    _vault_cleanup_frontmatter, _vault_tag_migrate, _vault_rebuild_overviews,
    _vault_polish_markdown, _bases_emit
  - cli_pipeline.py (13): _cmd_doctor, _cmd_ingest, _fit_check_emit/apply/
    audit/drift, _import_folder_command, _import_folder_dep_precheck, _auto,
    _sync_status, _sync_reconcile, _pipeline_repair, _migrate_yaml

cli.py re-exports all 23 (external importers + _main_dispatch unchanged);
6231 -> 5373 LOC.

Dependency-sync forwarders (module-level + _sync_cli_dependencies, mirroring
the existing cli_notebooklm/cli_zotero pattern) so test monkeypatches on
research_hub.cli.* still reach the moved handlers:
  - _cli_vault: get_config, ClusterRegistry
  - _cli_pipeline: get_config, ClusterRegistry, run_pipeline, repair_cluster,
    _load_zotero_if_configured

Consolidate triplicated _load_zotero_if_configured (was duplicated in
cli_clusters + cli_zotero, plus a stripped-docstring copy the delegate added
to cli_common) into ONE canonical def in cli_common with the full v0.90.0
G1#1 docstring retained; cli/cli_clusters/cli_pipeline now import it from
cli_common (no sibling cli_* imports).

Tests unchanged (pure move). Full suite 3144 passed, 0 failed.
Extraction delegated to Codex; cli.py surgery + helper consolidation by Claude.

Co-Reviewed-By: code-reviewer subagent (triggers 1+4: >=3 files / delegate-returned) -> APPROVE
