+++ src/research_hub/cli_zotero.py
@@ -387,30 +387,3 @@ def _zotero_backfill(args) -> int:
     if args.apply and report.report_path:
         print(f"Markdown report saved: {report.report_path}")
     return 0
-
-def _load_zotero_if_configured():
-    """Lazy-load Zotero client. Returns None if not configured.
-
-    v0.90.0 G1#1 fix: distinguish "not configured" (silent None) from
-    "configured but broken" (warn to stderr, still return None). Pre-fix,
-    the bare ``except Exception`` made auth failures, network outages, and
-    missing imports all look identical to "no Zotero set up", so users
-    saw zero ingestion and assumed they hadn't configured Zotero when in
-    reality the client was broken.
-    """
-    try:
-        from research_hub.errors import MissingCredential
-        from research_hub.zotero.client import get_client
-
-        return get_client()
-    except MissingCredential:
-        # Truly unconfigured -- silent None preserves lazy-mode UX
-        return None
-    except Exception as exc:
-        # Configured but broken -- surface root cause so user can act
-        print(
-            f"  [zotero] WARN credentials present but client init failed: "
-            f"{type(exc).__name__}: {exc}",
-            file=sys.stderr,
-        )
-        return None
